@@ -7,4 +7,6 @@ with support for comments extraction and image downloading.
 
 __version__ = "0.2.0"
 __author__ = "sadadYes"
-__description__ = "A Python tool for archiving YouTube community posts with zero dependencies"
+__description__ = (
+    "A Python tool for archiving YouTube community posts with zero dependencies"
+)
